Repository refactoring
