+++ .github/workflows/ci_prepare.yml
@@ -28,9 +28,7 @@ jobs:
             - name: Lint lib src
               run: |
                 # stop the build if there are Python syntax errors or undefined names
-                flake8 temporal_boost --show-source --statistics
-                # exit-zero treats all errors as warnings
-                flake8 temporal_boost --count --exit-zero --statistics
+                ruff check .
             - name: Lint unit tests
               run: |
                 # stop the build if there are Python syntax errors or undefined names
